fix: nav items & nav pop over in seperate file
